requirements.txt
requests>=2.22.0
flask>=1.1.1
gunicorn>=20.0.2
python_dotenv>=0.10.3
scikit-learn>=0.21.3
matplotlib>=3.1.2
numpy>=1.17.4
joblib>=0.14.0
timeloop>=1.0.2
elasticsearch>=7.1.0
elasticsearch-dsl>=7.1.0
pandas>=0.25.3
uuid>=1.30